#835 Download ChromeDriver executable (#842)
* #835 Download ChromeDriver executable before execution
