+++ src/main/java/com/appium/manager/LocalAppiumManager.java
@@ -11,6 +11,7 @@ import com.github.iOS.IOSManager;
 import io.appium.java_client.service.local.AppiumDriverLocalService;
 import io.appium.java_client.service.local.AppiumServiceBuilder;
 import io.appium.java_client.service.local.flags.GeneralServerFlag;
+import io.appium.java_client.service.local.flags.ServerArgument;
 
 import java.io.File;
 import java.io.IOException;
